requirements.txt
alembic==0.9.5
autoenv==1.0.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-API==1.0
<<<<<<< HEAD
Flask-Bcrypt==0.7.1
=======
>>>>>>> ft-crud-api-funtionality-151246191
Flask-HTTPAuth==3.2.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
<<<<<<< HEAD
gunicorn==19.7.1
=======
>>>>>>> ft-crud-api-funtionality-151246191
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
